N9 X Clipboard: Fix for aegis block
Dereferencing the /proc/%d/exe symlink worked over SSH, but not directly
from the device... The cmdline file is readable on both, so pull the
executable name out of that (which relies on the clipboard manager being
involked with it's full path).

Ideally I guess it would be better to get the PID of the process
connected to X that sent the request. Not sure how to do that though...

One additional problem that I've noticed is that stopping the clipboard
manager seems cause a bunch of delays everywhere, which could be
problematic.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>