Blacklist sending credentials to qtcreator
If qtcreator is running, the moment we take ownership of an X selection
it requests it, so add it to the blacklist.

I am getting a strong suspicion that QT is in general doing the wrong
with wrt ICCCM given the nonsense that I've seen on the N9 as well.

I'll need to test more QT applications and see how deep this problem
really lies.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>