Don't try to delete non-existing fields
If a new http script is created, but cleared before saving a record it
would attempt to delete the empty field from the record, which would
raise an exception as that field did not yet exist.

Only attempt to delete an empty field if it actually exists in the
record.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>