X Clipboard: Partially blacklist TightVNC viewer
The moment the vnc viewer window gets focus it requests the PRIMARY
selection. Or steals it. Either way it is quite broken, especially as it
doesn't even seem to do anything with the selection at that point!

Add code to blacklist by a regular expression of the window title, since
the title of TightVNC includes the hostname it connected to so I can't
do an exact match.

The reason this is only "partially" blacklisted, is that using the F8
menu to transfer the local clipboard to the remote machine actually does
the right thing, and uses the popup window's name ("popup") to request
the selection...   Unfortunately this only works if vncviewer did not
steal the PRIMARY selection from us.

A full workaround will likely involve re-grabbing the selection whenever
TightVNC attempts to steal it (this might be useful for certain other
applications as well), however I'm reluctant to do this in the general
case as it leads to users unable to use the clipboard themselves until
kosh is done (on the other hand, if kosh is waiting to send something,
they clearly requested it...).