Force config file updates to use utf-8
This has been working perfectly for years then today it mysteriously
wrote the updated password in utf-32. We aren't handling unicode at all
yet so force it to write out utf-8 to be compatible with ascii.

Really the project needs to be ported to Python 3 so any other potential
unicode bugs get found and fixed.