Add basic clipboard integration on Windows / cygwin
This adds basic support for the Windows clipboard. It currently sets the
clipboard to each yankable blob in turn, requiring the user to press enter in
the kosh window between them.

It might be possible to add redimentary detection for an application pasting
the clipboard in the future by setting the clipboard data handle to NULL and
rendering it on demand with WM_RENDERFORMAT / WM_RENDERALLFORMATS, however the
Windows API does not appear to provide any sort of notification once the
application has finished pasting the data, meaning we might need to delay a
short time and hope for the best. The Windows clipboard API also appears to
lack notification of which application is requesting the clipboard, meaning
blacklisting bad / misbehaving applications might be impossible.