Windows Clipboard: Automatically advance field when clipboard pasted
This implements an approximation of the clipboard functionality we have
under X to automatically advance from copying the username to the
password when an application pastes the clipboard.

To achieve this we create a hidden window that takes ownership of the
clipboard, and indicates that it has text to paste in a deferred manner
(i.e. passes NULL for the handle). When the target application pastes
the clipboard our window is sent a WM_RENDERFORMAT message and we can
place the current field in the clipboard at that time.

Unlike X, there is nothing in the API we can use to determine when the
receiving application has finished obtaining data from the clipboard, so
we start a short timer after setting the clipboard to empty the
clipboard and advance to the next field.

Since having a win32 Window on screen requires that we use the win32
message queue we once again find ourselves outside of the urwid main
loop but still need to process it. We currently use a hack to call
select() on the UI file descriptors periodically from a Windows timer to
process this, but later should split the win32 window and message queue
out into a separate thread.

While Python can recognise that it is running under cygwin, the ctypes
module does not load full Windows support in this case. However, since
this is a hard requirement for me, this includes a bunch of
compatibility code to fill out enough of the missing wintypes module to
support all the clipboard APIs.