Clipboard/Windows: Add a capital yank keybinding without auto-advance
This adds an alternate key binding to copy the selected record/field to
the clipboard without the usual auto-advance when an application
requests the clipboard, instead requiring that the user press
enter/escape when ready. This can useful for copying a password to
both the password and confirm password fields when creating a new
account, and potentially for dealing with any problematic applications
that don't follow typical clipboard behaviour.

This is currently only implemented on Windows and simply uses the
previous simpler version of the clipboard code that just sets the
clipboard without creating an owning window.