Prevent password list from being focussed while editing an entry
If the password list is inadvertently focussed while editing an entry, a
different entry may be selected which will wipe out the current edit in
progress, which is bad. Make the password list unselectable while
editing an entry to prevent this.

Also fix some issues with selection via the mouse - it can no longer
switch to revealing an entry while editing it (which would have also
lost the edits), and the first press now only shows the password entry
without revealing it, requiring a second press to reveal it.