Lock program after timeout
This adds a countdown timer to the UI that once expired will terminate
the program provided that no edit is in progress. Pressing any key will
restart the countdown timer.

If an edit is in progress when the timer expires it will defer
terminating the program to give the user a chance to finish their edit.
It will however hide the list of password entries and will exit once the
user has saved or cancelled.

Originally this functionality was going to go in the database module,
however driving it from the interface makes things considerably simpler.