Disable ptrace to mitigate opportunistic gcore attacks
This prevents the opportunistic attack whereby someone notices kosh left
running on an unattended workstation and runs gcore to extract passwords
from memory while the owner is distracted. Consider settings such as
university computer labs where this sort of attack may take place.

This will also boot out an already attached debugger, but this is *NOT*
trying to mitigate any attacks involving that avenue, since an existing
debugger could easily skip over the prctl call and there's nothing we
can really do to stop it. However, such an attack requires advance
preparation - it is not opportunistic, and if your threat model includes
such an adversary than your only defence is to ensure that they never
gain access your workstation to prepare this or any other attack in the
first place.