Add support for Android via termux-api
Termux only provides a very basic clipboard API compared that has no
mechanism to know when an application has pasted the contents or block
unwanted applications from reading the clipboard. We might want to
investigate bypassing termux later and talking straight to Android (that
is, assuming Android has a better API, which I wouldn't count on).