Switch to pycryptodome
pycrypto is no longer maintained, and recommended to switch to a fork
such as pycryptodome.

pycryptodome has changed the AES.new() function signature to require
passing in a 'mode' - to retain compatibility with existing databases
the previous default of ECB is used, however it sounds like this default
was deprecated for a good reason and we should consider changing.