Add basic support for clipboard on WSL (via clip.exe)
This adds rudimentary support for using the clipboard on WSL, by calling
out to clip.exe.

In theory updated versions of WSL 2 on Windows 11 (not 10) with the
appropriate WSL graphics drivers installed can now use Xlib and the
clipboard is supposed to work through that - and if that worked we might
possibly be able to support more of the full rich clipboard experience
that we have under Linux. However, I have my doubts that the advanced
functionality would work as it doesn't have a good analogue on Windows
API to translate to, and in practice even basic support doesn't seem to
work at all.