WSL Clipboard: Call out to native Python to support advanced clipboard
WSL clipboard via xlib/wayland just doesn't work for me, and the callout
to clip.exe we implemented as a workaround is also too basic to know
when the user has pasted the contents.

Instead, call out to the old winclipboard.py via native windows python
to implement this, so we at least get parity with the native windows
implementation. This is only a rudimentary implementation of this, and
doesn't yet give status updates, nor processes enter/escape keys.