WSL: Implement status, enter to skip, escape to cancel functionality
This moves the blob iteration from winclipboard to wslclipboard so that
it can use a select() loop to respond to user input, and to detect when
winclipboard terminates by noting when it's stdout has no more data.

Practically, this makes wslclipboard just as usable as winclipboard,
making kosh viable in WSL.

Currently, it will not report messages from winclipboard as that would
lead to duplicate messages with confusing timing. The downside of this
is that we don't see the reports of ignoring clipboard pastes within x
ms of taking the clipboard.