Update revealed TOTP codes live, add re-hide functionality
If a TOTP code is revealed, show a countdown until it expires and update
to show the new code when the timer reaches 0.

This also addresses the long standing issues where a revealed field
could no longer be selected or hidden again without changing to a
different entry and back again. Revealed fields are now displayed as
buttons that can be activated to re-hide them.