Keep WSL clipboard proxy running
Launching the clipboard proxy each time we want to send a new entry can
be slow, especially if something like Crowdstrike wants to feel needed.

Rather than launching winclipboard in native python as a proxy every
time wslclipboard wants to send some data, launch it once on init and
keep it running until kosh is closed, sending new entries to copy to the
clipboard via stdin, as well as special sequences to signal when the
user has pressed enter to skip or escape to cancel to the current
record.

This retains the Windows Defender warning if the proxy takes an
excessive amount of time to launch, because when that was observed the
entire execution was slow, not just the launch, so the user may still
wish to add an exception.

IPC is always a bit of a PITA to get right when using OS pipes, and has
been no exception here - so this may need further testing.