Fix launching on native Windows
Probably a long time since I mainly run this under a Unix environment,
even on Windows, so bit rot has broken it in a number of ways.

disable_ptrace() doesn't work on Windows, so make that platform
specific. There is an equivalent on Windows that I'll need to track
down.

Remove import of fcntl that does not exist on Windows, and implement an
alternative file locking scheme.

Close files before renaming them, and remove existing files before
renaming over the top.