Work around an issue where the first ever entry can't be edited
Testing with a brand new password database there seems to have been a
regression preventing the very first entry from being edited. Not
entirely sure what's going on here, but it seems that urwid doesn't
think the password form is selectable for some reason... So for now make
it always selectable to work around the issue.