Fix clipboard input handling on native Windows
Window's select() implementation doesn't support pipes like stdin, so
urwid uses a thread to pass stdin to a socket that is supported by
select. Because of this, on Windows we can't read from stdin in the
clipboard input handler, so add a check for the urwid socket file number
and recv from it instead.