Fix stalls/errors with clipboard after WSL clipboard proxy has emptied clipboard
Turns out to empty the clipboard we should not be passing the hWnd, as
doing so will still result in the proxy window becoming the clipboard
owner, which causes stalls and/or errors when another program attempts
to take the clipboard later. This is apparent by using a regular yank
followed by a capital yank - the capital yank will take much longer than
normal.

This has only become a problem since switching the clipboard proxy to
remain running, as previously we would destroy the clipboard window once
we were done which would implicitly release the clipboard ownership.