Port clipboard input fix to termux + other platforms
Android / Termux has started exhibiting the same bug that we saw on
Windows where input would be passed through to urwid when it was
supposed to be consumed by the clipboard input loop. Port the fix
checking for the name of an io.TextIOWrapper to all other clipboard
implementations. Don't port the socket path, since that is specific to
win32.