Implement QR code import from clipboard on WSL/Windows
This uses OpenCV to decode a QR code in the clipboard and import the
value as a field in a password record, mostly intended to import TOTP
provisioning URLs embedded in QR codes, i.e. as an alternative to using
authenticator apps.

Activating the new field while editing a password record will try to
decode a QR Code from the clipboard, and if successful will add it's
contents (usually a provisioning URL) to a new field (named "2FA" by
default) in the password record. The normal workflow would then be to
save the record and kosh's existing TOTP handler will allow the OTP code
to be copied to the clipboard or displayed in order to confirm to the
provisioning website that the code has been imported successfully.

This is currently only implemented on WSL/Windows, though besides the
WSL->native Windows proxy most of the code should be platform agnostic
(PIL handles the platform specific aspect of getting the clipboard
contents as an image), so supporting it on other platforms should be
pretty trivial.

Note that this does not handle the QR codes exported from Authenticator
apps, as they contain many provisioning URLs encoded in an additional
layer, and often need to use multiple QR codes to send all entries!
Supporting that within the app will need some thought as to how to
distribute the 2FA codes to different records, handle previously
imported codes, etc.