Fix crash if y/Y/q is pressed on save/cancel buttons
These key bindings are enabled in edit mode, though usually don't get
triggered as the text fields will consume the inputs first. However, if
some other widget is highlighted, such as the save/cancel buttons, the
input is not consumed and the callbacks can be triggered, causing a
crash since the selected widget is not a field in a password record.

Disable them all in edit mode to prevent the crash. Notably this fix was
already done to the 'S' key binding.