Prevent editing/deleting a record while editing + fix delete
The left hand password list is not intended to be selectable while
editing a record, but seems that it is anyway. Make sure that edit and
delete cannot be used while already editing another entry as that seems
like it would be bad.

Other functions seem relatively safe to use in this state even if they
aren't intended to be usable, so don't worry too much about those.

This also fixes a regression where deleting an entry doesn't update the
password list since it didn't get removed from the visible entries, and
adds a yes/no confirmation dialog when deleting an entry.