wslclipboard: Use python3.exe to launch native Python
python.exe may be either Python 2 or 3, so shouldn't rely on this since
we specifically want Python 3. Change the executable to python3.exe,
which will hopefully do what we want.

Still need to double check what python3.exe does if it is not installed
- is it a Windows Store stub like python.exe, or does it just not exist?
For now left the stub and installer checks to python.exe, which should
work either way.